# Touride/src/Touride/src/Touride.UI/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "sample",
    "description": "sample",
    "contact": {
      "name": "API"
    },
    "license": {
      "name": " TG Labs License"
    },
    "version": "v1"
  },
  "paths": {
    "/api/v1/Notification/emailsend": {
      "post": {
        "tags": [
          "Notification"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailMessageDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailMessageDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmailMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/api/v1/Sample/insert": {
      "post": {
        "tags": [
          "Sample"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/api/v1/Sample/getall": {
      "get": {
        "tags": [
          "Sample"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/api/v1/Sample/getalldevextreme": {
      "post": {
        "tags": [
          "Sample"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataSourceLoadOptions"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataSourceLoadOptions"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataSourceLoadOptions"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/api/v1/SampleEvent/TestIntegrationEvent": {
      "post": {
        "tags": [
          "SampleEvent"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestIntegrationEvent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TestIntegrationEvent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TestIntegrationEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AddressDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateTestModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataSourceLoadOptions": {
        "type": "object",
        "properties": {
          "requireTotalCount": {
            "type": "boolean"
          },
          "requireGroupCount": {
            "type": "boolean"
          },
          "isCountQuery": {
            "type": "boolean"
          },
          "isSummaryQuery": {
            "type": "boolean"
          },
          "skip": {
            "type": "integer",
            "format": "int32"
          },
          "take": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortingInfo"
            },
            "nullable": true
          },
          "group": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupingInfo"
            },
            "nullable": true
          },
          "filter": {
            "type": "array",
            "items": {},
            "nullable": true
          },
          "totalSummary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SummaryInfo"
            },
            "nullable": true
          },
          "groupSummary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SummaryInfo"
            },
            "nullable": true
          },
          "select": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "preSelect": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "remoteSelect": {
            "type": "boolean",
            "nullable": true
          },
          "remoteGrouping": {
            "type": "boolean",
            "nullable": true
          },
          "expandLinqSumType": {
            "type": "boolean",
            "nullable": true
          },
          "primaryKey": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "defaultSort": {
            "type": "string",
            "nullable": true
          },
          "stringToLower": {
            "type": "boolean",
            "nullable": true
          },
          "paginateViaPrimaryKey": {
            "type": "boolean",
            "nullable": true
          },
          "sortByPrimaryKey": {
            "type": "boolean",
            "nullable": true
          },
          "allowAsyncOverSync": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "EmailMessageDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "nullable": true
          },
          "tos": {
            "type": "string",
            "nullable": true
          },
          "cCs": {
            "type": "string",
            "nullable": true
          },
          "bcCs": {
            "type": "string",
            "nullable": true
          },
          "subject": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GroupingInfo": {
        "type": "object",
        "properties": {
          "selector": {
            "type": "string",
            "nullable": true
          },
          "desc": {
            "type": "boolean"
          },
          "groupInterval": {
            "type": "string",
            "nullable": true
          },
          "isExpanded": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LoadResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {},
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "default": -1
          },
          "groupCount": {
            "type": "integer",
            "format": "int32",
            "default": -1
          },
          "summary": {
            "type": "array",
            "items": {},
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "SortingInfo": {
        "type": "object",
        "properties": {
          "selector": {
            "type": "string",
            "nullable": true
          },
          "desc": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SummaryInfo": {
        "type": "object",
        "properties": {
          "selector": {
            "type": "string",
            "nullable": true
          },
          "summaryType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TestIntegrationEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "description": "Balea Server OpenId Security Scheme",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://authorizationurl.domain.com/connect/authorize",
            "scopes": {
              "test-write": "test-write",
              "test-read": "test-read"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "Bearer": [
        "openid",
        "profile",
        "email"
      ]
    }
  ]
}